pyproject.toml
[tool.poetry]
name = "chatgpt-response"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "chatgpt_response"}]

[tool.poetry.dependencies]
python = "^3.9.0"
flask-unittest = "^0.1.3"
fluentcheck = "^0.2.0"
requests = "^2.28.2"
dependency-injector = "^4.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
